chore: move tests to contract folder
